refactor: extract project-filesystem.js from project.js (PR-10)
Move file browser handlers (fs_list, fs_read, fs_write), project
settings/env handlers, file watcher delegation, file edit history,
and git diff/file-at handlers into a dedicated module. project.js
reduced from 3,278 to 2,878 lines.
